+++ samples/asynctests/test_azure_iothub_send_async.py
@@ -32,7 +32,7 @@ def get_logger(level):
     return uamqp_logger
 
 
-log = get_logger(logging.DEBUG)
+log = get_logger(logging.INFO)
 
 
 def _generate_sas_token(uri, policy, key, expiry=None):
@@ -75,7 +75,7 @@ async def test_iot_hub_send_async(live_iothub_config):
     target = 'amqps://' + endpoint + operation
     log.info("Target: {}".format(target))
 
-    send_client = uamqp.SendClientAsync(target, debug=True)
+    send_client = uamqp.SendClientAsync(target, debug=False)
     send_client.queue_message(message)
     results = await send_client.send_all_messages_async()
     assert not [m for m in results if m == uamqp.constants.MessageState.SendFailed]