+++ uamqp/async_ops/session_async.py
@@ -4,10 +4,10 @@
 # license information.
 #--------------------------------------------------------------------------
 
-import asyncio
 import logging
 
 from uamqp import constants, errors, session
+from uamqp.utils import get_running_loop
 from uamqp.async_ops.mgmt_operation_async import MgmtOperationAsync
 
 _logger = logging.getLogger(__name__)
@@ -46,7 +46,7 @@ class SessionAsync(session.Session):
                  handle_max=None,
                  on_attach=None,
                  loop=None):
-        self.loop = loop or asyncio.get_event_loop()
+        self.loop = loop or get_running_loop()
         super(SessionAsync, self).__init__(
             connection,
             incoming_window=incoming_window,
@@ -81,7 +81,7 @@ class SessionAsync(session.Session):
         :type node: bytes or str
         :param timeout: Provide an optional timeout in milliseconds within which a response
          to the management request must be received.
-        :type timeout: int
+        :type timeout: float
         :param status_code_field: Provide an alternate name for the status code in the
          response body which can vary between services due to the spec still being in draft.
          The default is `b"statusCode"`.